Dedicate a module for export
